requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pillow==10.1.0
greenlet==3.0.1
redis==5.0.1